canister_tests.rs file was getting too big, splitting it up
