+++ bin/icp-cli/tests/canister_stop_tests.rs
@@ -1,0 +1,78 @@
+use crate::common::TestEnv;
+use icp_fs::fs::write;
+use predicates::{
+    prelude::PredicateBooleanExt,
+    str::{contains, starts_with},
+};
+use serial_test::serial;
+
+mod common;
+
+#[test]
+#[serial]
+fn canister_stop() {
+    let env = TestEnv::new().with_dfx();
+
+    // Setup project
+    let project_dir = env.create_project_dir("icp");
+
+    // Use vendored WASM
+    let wasm = env.make_asset("example_icp_mo.wasm");
+
+    // Project manifest
+    let pm = format!(
+        r#"
+        canister:
+          name: my-canister
+          build:
+            adapter:
+              type: script
+              command: echo {}
+        "#,
+        wasm,
+    );
+
+    write(
+        project_dir.join("icp.yaml"), // path
+        pm,                           // contents
+    )
+    .expect("failed to write project manifest");
+
+    // Start network
+    let _g = env.start_network_in(&project_dir);
+
+    // Wait for network
+    env.configure_dfx_local_network();
+
+    env.dfx()
+        .arg("ping")
+        .arg("--wait-healthy")
+        .assert()
+        .success();
+
+    // Deploy project
+    env.icp()
+        .current_dir(&project_dir)
+        .args(["deploy", "--effective-id", "ghsi2-tqaaa-aaaan-aaaca-cai"])
+        .assert()
+        .success();
+
+    // Stop canister
+    env.icp()
+        .current_dir(&project_dir)
+        .args(["canister", "stop", "my-canister"])
+        .assert()
+        .success();
+
+    // Query status
+    env.icp()
+        .current_dir(&project_dir)
+        .args(["canister", "status", "my-canister"])
+        .assert()
+        .success()
+        .stderr(
+            starts_with("Canister Status Report:")
+                .and(contains("Status: Stopped"))
+                .and(contains("Controllers: 2vxsx-fae")),
+        );
+}
